pyproject.toml
[tool.poetry]
name = "employee-lunch"
version = "0.1.0"
description = ""
authors = ["Bezsmertnyi"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"
psycopg2-binary = "^2.9.9"
pyjwt = "^2.8.0"
setuptools = "^70.1.0"
python-decouple = "^3.8"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
lint = "flake8"